Migrate a batch of table components
